+++ Dos.Game/GameConfig.cs
@@ -9,12 +9,14 @@ namespace Dos.Game
             FalseCalloutPenalty = 2;
             CalloutPenalty = 2;
             StartingPlayer = null;
+            DoubleColorMatchDraw = 1;
         }
 
         public ushort CalloutPenalty { get; set; }
         public ushort FalseCalloutPenalty { get; set; }
         public ushort InitialHandSize { get; set; }
         public ushort MinCenterRowSize { get; set; }
+        public ushort DoubleColorMatchDraw { get; set; }
         public bool CenterRowPenalty { get; set; }
         public bool DrawEndsTurn { get; set; }
         public bool SevenSwap { get; set; }