+++ changelog.md
@@ -4,9 +4,16 @@
  - Added configurable rules: decks amount, penalties for callout and false callout, center row size, use images or text for displaying cards and first two house rules
  - Added **Center Row Penalty** house rule: at the end of your turn, draw as much cards as you left unmatched in the center row
  - Added **Draw Ends Turn** house rule: if you draw a card, your turn ends immediately
- 
+
  #### v1.1.1 (23.01.2020)
  - Fixed counting match of two ones of different colors with Wild 2 as Double Color Match
- 
+
  #### v1.1.2 (25.01.2020)
  - Added some clarifications in help command
+
+ #### v1.2.0 (08.03.2020)
+ - Now quitting does not stop the game as long as there's more that 1 player
+ - Improved game logging
+ - Added **Seven Swap** house rule: Color Match on 7 will force you to switch your hand  with somebody else
+ - Added config option to change amount of cards players have to draw when somebody else made a Double Color Match 
+ - Added bot players: right now they make random moves